refactor: deleted export as songsModel
