+++ src/entity/songs/index.ts
@@ -1,3 +1,1 @@
-export * as songModel from "./model";
-
 export * from "./model";
