+++ src/features/add-to-favorites/model.ts
@@ -1,9 +1,9 @@
 import { createEvent, sample } from "effector";
-import { songModel } from "~/entity/songs";
+import { addToFavoriteButtonClicked } from "~/entity/songs";
 
 export const addTofavoritesButtonClicked = createEvent<number>();
 
 sample({
   clock: addTofavoritesButtonClicked,
-  target: songModel.addToFavoriteButtonClicked,
+  target: addToFavoriteButtonClicked,
 });
