+++ src/features/music-page/search-input/view.tsx
